feat(aihub_web): Merge expert questions and groups UIs into unified experts service
- Create unified /service/expert page with TabView (Questions + Groups tabs)
- Add tab labels to i18n (de, en, fr, it)
- Move ExpertGroup components to Expert directory with Group prefix
- Update group detail page close route to return to unified page
- Remove old separate questions.vue and groups.vue pages
- Add URL query param support for deep linking (?tab=questions|groups)
- Display pending questions badge on Questions tab header
