+++ aihub_web/aihub_web/i18n/locales/en.yaml
@@ -540,6 +540,9 @@ process:
 
 expert:
   title: "Experts"
+  tabs:
+    questions: "Questions"
+    groups: "Groups"
   pending_questions: "Pending Questions"
   no_pending_questions: "No pending questions"
   no_answered_questions: "No answered questions"
