+++ aihub_web/aihub_web/i18n/locales/it.yaml
@@ -517,6 +517,9 @@ process:
 
 expert:
   title: "Esperti"
+  tabs:
+    questions: "Domande"
+    groups: "Gruppi"
   pending_questions: "Domande in sospeso"
   no_pending_questions: "Nessuna domanda in sospeso"
   no_answered_questions: "Nessuna domanda con risposta"
